Base of Indoor Tileset
So this is the base of the tileset, theres a sample too. If a different
color other than that creamish off white let me know and it shouldn't
take long to change. I am working on the windows and a couple of other
things to put on the walls that I will send when i finish it.