Maps And Stuff
Started to work on the outside fighting when I realized there were going
to be issues with walking behind buildings and whatnot. Started the
remake process with the entry square, buildings at different depths now
need to have their own layers and shadows and all that need to be remade
due to the transparency. The other addition is that of hitboxes, as also
seen in the entry square.