Factory Map
Worked on the factory map a lot more at home, helped tommy fix the
corners and stuff as needed for the map. Then I planned out an addition
to the map where the player fights a miniboss to turn on the factory and
has to go into the back conveyor room to get around an obstacle blocking
the rest of the factory.