Factory
I added more sprites to the factory and learned how to do so in tiled,
takes a lot of time to line them all up right. Also worked on making
sure the map was actually finished, checked the hitboxes and had to
figure out what was blocking the rest of the building off for a while.