OutsideFighting Map
The outside fighting map is almost done, it has been remade yet again to
have smaller buildings along the conveyor belts. All the hitboxes are
being moved around and the factory still has to be added. This is
continued from my work last night. I also added the new sprites and
worked on testing the factory a bit last night.